pyproject.toml
[project]
name = "advanced-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.6",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
